Heroku deployments set up
Heroku deployments have now been hooked into the github repo. This
branch will be merged into trunk and then subsequently a deployment will
occur with the latest build off trunk. This branch will be deleted, the
only need for it was because a manual heroku CLI deployment requires a
branch named either master or main. Having both trunk and main is an
unnecessary confusion.